test: Reorganize NDV tests into workflows/editor/ndv/ directory
