+++ packages/testing/playwright/tests/ui/workflows/editor/ndv/io-filter.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../../fixtures/base';
+import { test, expect } from '../../../../../fixtures/base';
 
 test.describe('Node IO Filter', () => {
 	test.beforeEach(async ({ n8n }) => {
